Trigger "changeDirectory" even on URL change
When using the browser back button or clicking on sections on the left
sidebar (like favorites), the "changeDirectory" jQuery event did not get
called, so apps like recommendations would not notice the directory
change.

This fixes the issue by also setting changeUrl to true when the file
list's directory got changed as a result from a URL change.

Note that in the past this argument was there to prevent repushing the
same URL to the history stack but is obsolete already as there's a check
in place to not push again whatever is already there.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>